+++ packages/blinkcard-ux-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @microblink/blinkcard-ux-manager
 
+## 3000.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/blinkcard-core@3000.0.6
+
 ## 3000.0.5
 
 ### Patch Changes